Restructure schema locations.
* Move schemas from config and context into their own files under the
  antsibull/schemas directory.
* Move doc parsing related schemas to antsibull/schemas/docs.
* Add a deprecated antsibull.schemas.ansible_doc module for backwards
  compat.  The new location is antsibull.schemas.docs.ansible_doc
* Extract validators for converting booleans and None from strings to a
  schemas/validators.py file.  These were used by both config and
  context so it makes sense to put them in one place.
* Move ContextDict into utils.collections to avoid circular import.
  The app_context.py code requires the context schemas and the context
  schemas require the ContextDict data structure.  ContextDict was
  living in app_context.py which would lead to a circular import.
  Moving it to utils/collections.py breaks the cycle.
* Update documentation and tests for the new location of docs schemas
