+++ antsibull/schemas/docs/plugin.py
